% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_fill_mn_state.R
\name{scale_fill_mn_state}
\alias{scale_fill_mn_state}
\title{Minnesota state colors for fill aesthetics in ggplot}
\usage{
scale_fill_mn_state(
  palette = "primary brand",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{The color palette to use}

\item{discrete}{Boolean for whether the data is discrete}

\item{reverse}{Boolean to revers the order of the categories}

\item{...}{Other parameters passed to ggplot2::discrete_scale() or ggplot2::scale_color_gradientn()}
}
\value{
ggplot2 object with Minnesota state branding colors
}
\description{
Minnesota state colors for fill aesthetics in ggplot
}
